Pass collection and slug through to OG image template
Did so by joining them with a slash. This has the added benefit of
ensuring that they have unique URLs
